[Bronze I] Title: 더하기 2, Time: 44 ms, Memory: 31120 KB -BaekjoonHub

+++ 백준/Bronze/10823. 더하기 2/더하기 2.py
@@ -1,1 +1,1 @@
-print(int(sum(map(float,open(0).read().replace('\n','').split(',')))))
+print(sum(map(int,open(0).read().replace('\n','').split(','))))
